feat: add `agent-dispute-closure`
New methods allow users to register an agent that is able to dispute
closures in their stead. This allows users to separate the key that is
able to spend from their channels and another key that is able to
dispute channel closures. This agent key can be kept live on a server
which monitors the channels.