feat: allow agent to sign all responsive messages
This means that an agent A for principal B cannot sign off on a transfer
from B to C, but it can sign off on a transfer from C to B. This allows
the agent to handle all activity needed to manage an account, but does
not allow it to initiate actions on behalf of the user.

It is important to note that this enables the agent to steal funds if
they are in collaboration with the other party of a channel.