feat: add `valid-after` parameter to transfers
This new parameter allows users to sign a transfer that is optionally
only valid after the specified burn block height has passed. This can be
useful in various scenarios where a user may need to pre-sign a transfer
but not allow it to be used until a specific block height.