feat: simplify receive liquidity
In this new model, the Reservoir operator is the one and only provider
of liquidity. It makes more sense to do this, since having other
liquidity providers requires the LPs to trust the Reservoir, so if they
are going to trust the Reservoir, then we may as well move the details
off-chain and simplify the contract.